A basic messenger made with Python 2.7
This is a project I am making for my 12th Grade CBSE examinations.
This messenger works best with on Computers woth the same network. (I
haven't tested it with the Internet)
This is the alpha version of PyMessenger.
More features will be added soon.
Also the UI will be improved.( It will still be Command Line. :P )
If I get time to make it GUI ( using TKinter ), I will. But that is way
down the priority list. :P
Testers are welcome.